Further work on the conversation settings screens
Added some missing strings.
Added in a simple data binding pattern and updated the settings screens to use it (will see how it goes).
Added in the updated notification settings detail screen for groups.
Updated the layout to be a little more responsive for small devices.
Updated the ConversationSettingsActionView to have a disabled state (for after the user has left a group).