Added additional functionality to conversation settings
Added the ability to view the profile picture.
Added the ability to invite a user to join an open group.
Added the ability to edit the name of a note to self or DM conversation.
Minor tweaks and optimisations for the data binding classes.
Made a few changes to try and clean up the ConversationSettingsViewModel and how interacting with it works.